refactor(card6/parte2) : change the type of product dimensions
fixes #card6 - parte 2 - Deve calcular o valor do frete com base nas dimensões (altura, largura e profundidade em cm) e o peso dos produtos (em kg)

+++ src/main/java/br/com/branas/glstore/domain/entities/Product.java
@@ -14,9 +14,9 @@ public class Product {
 
     private String productDescription;
     private BigDecimal productPrice;
-    private Integer productWidth;
-    private Integer productHeight;
-    private Integer productLength;
+    private Double productWidth;
+    private Double productHeight;
+    private Double productLength;
     private Double productWeight;
 
     public Product() {
@@ -43,27 +43,27 @@ public class Product {
         this.productPrice = precoProduto;
     }
 
-    public Integer getProductWidth() {
+    public Double getProductWidth() {
         return productWidth;
     }
 
-    public void setProductWidth(Integer productWidth) {
+    public void setProductWidth(Double productWidth) {
         this.productWidth = productWidth;
     }
 
-    public Integer getProductHeight() {
+    public Double getProductHeight() {
         return productHeight;
     }
 
-    public void setProductHeight(Integer productHeight) {
+    public void setProductHeight(Double productHeight) {
         this.productHeight = productHeight;
     }
 
-    public Integer getProductLength() {
+    public Double getProductLength() {
         return productLength;
     }
 
-    public void setProductLength(Integer productLength) {
+    public void setProductLength(Double productLength) {
         this.productLength = productLength;
     }
